+++ tf_yarn/client.py
@@ -1,4 +1,3 @@
-
 import importlib
 import logging
 import logging.config
