+++ tf_yarn/tensorflow/tasks/_tensorboard_task.py
@@ -1,13 +1,11 @@
 import logging
 import os
-import skein
-import tensorflow as tf
 
-from typing import Optional
+import skein
 
 from tf_yarn import Experiment, KerasExperiment
-from tf_yarn.tasks import logging as tf_yarn_logging
-tf_yarn_logging.setup()
+from tf_yarn._task_commons import setup_logging
+setup_logging()
 
 from tf_yarn import (
     _task_commons,
