+++ tf_yarn/tensorflow/tasks/evaluator_task.py
@@ -7,9 +7,9 @@ import skein
 import tensorflow as tf
 
 from tf_yarn import _task_commons, event, cluster, Experiment, KerasExperiment
-from tf_yarn.tasks import logging as tf_yarn_logging
 from tf_yarn.tensorflow import metrics
-tf_yarn_logging.setup()
+from tf_yarn._task_commons import setup_logging
+setup_logging()
 
 logger = logging.getLogger(__name__)
